move Scale* and Translate to legacy, fix issues
