+++ legacy/core/src/main/java/org/eclipse/imagen/media/opimage/ScaleGeneralOpImage.java
