+++ legacy/core/src/main/java/org/eclipse/imagen/operator/ScaleDescriptor.java
