+++ legacy/core/src/main/resources/org.eclipse.imagen/org.eclipse.imagen.operator.properties
@@ -119,3 +119,17 @@ RescaleDescriptor1=The per-band constants to multiply by.
 RescaleDescriptor2=The per-band offsets to be added.
 RescaleDescriptor3=constants.length must be at least 1.
 RescaleDescriptor4=offsets.length must be at least 1.
+
+
+ScaleDescriptor0=Resizes an image.
+ScaleDescriptor1=The X scale factor.
+ScaleDescriptor2=The Y scale factor.
+ScaleDescriptor3=The X translation.
+ScaleDescriptor4=The Y translation.
+ScaleDescriptor5=The interpolation method for resampling.
+ScaleDescriptor6= operation requires both scale factors be greater than 0.
+
+TranslateDescriptor0=Moves an image to a new location.
+TranslateDescriptor1=The displacement in X direction.
+TranslateDescriptor2=The displacement in Y direction.
+TranslateDescriptor3=The interpolation method.
