+++ modules/bandmerge/src/test/java/org/eclipse/imagen/media/bandmerge/BandMergeTest.java
@@ -47,8 +47,8 @@ import org.eclipse.imagen.media.iterators.RandomIterFactory;
 import org.eclipse.imagen.media.range.Range;
 import org.eclipse.imagen.media.range.RangeFactory;
 import org.eclipse.imagen.media.testclasses.TestBase;
+import org.eclipse.imagen.media.translate.TranslateDescriptor;
 import org.eclipse.imagen.media.util.ImageUtil;
-import org.eclipse.imagen.operator.TranslateDescriptor;
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
