+++ modules/core/src/main/resources/org.eclipse.imagen/org.eclipse.imagen.operator.properties
@@ -395,13 +395,13 @@ RGBToIHSDescriptor0=Convert an image from RGB to IHS color space.
 #RotateDescriptor3=The rotation angle in radians.
 #RotateDescriptor4=The interpolation method.
 #RotateDescriptor5=The user-specified background values.
-ScaleDescriptor0=Resizes an image.
-ScaleDescriptor1=The X scale factor.
-ScaleDescriptor2=The Y scale factor.
-ScaleDescriptor3=The X translation.
-ScaleDescriptor4=The Y translation.
-ScaleDescriptor5=The interpolation method for resampling.
-ScaleDescriptor6= operation requires both scale factors be greater than 0.
+#ScaleDescriptor0=Resizes an image.
+#ScaleDescriptor1=The X scale factor.
+#ScaleDescriptor2=The Y scale factor.
+#ScaleDescriptor3=The X translation.
+#ScaleDescriptor4=The Y translation.
+#ScaleDescriptor5=The interpolation method for resampling.
+#ScaleDescriptor6= operation requires both scale factors be greater than 0.
 #ShearDescriptor0=Shears an image.
 #ShearDescriptor1=The shear value.
 #ShearDescriptor2=The shear direction.
@@ -439,10 +439,10 @@ ThresholdDescriptor6=operation requires all parameters to be of type double[].
 ThresholdDescriptor7=operation requires all parameters to have at least 1 array element.
 TransposeDescriptor0=Reflects an image in a specified direction or rotates an image in multiples of 90 degrees.
 TransposeDescriptor1=The The type of flip operation to be performed.
-TranslateDescriptor0=Moves an image to a new location.
-TranslateDescriptor1=The displacement in X direction.
-TranslateDescriptor2=The displacement in Y direction.
-TranslateDescriptor3=The interpolation method.
+#TranslateDescriptor0=Moves an image to a new location.
+#TranslateDescriptor1=The displacement in X direction.
+#TranslateDescriptor2=The displacement in Y direction.
+#TranslateDescriptor3=The interpolation method.
 UnsharpMaskDescriptor0=Performs UnsharpMask operation on the image.
 UnsharpMaskDescriptor1=The low-pass convolution kernel.
 UnsharpMaskDescriptor2=The sharpening value.
