+++ modules/mosaic/src/test/java/org/eclipse/imagen/media/mosaic/ComparisonTest.java
@@ -28,9 +28,9 @@ import org.eclipse.imagen.media.JAIExt;
 import org.eclipse.imagen.media.range.Range;
 import org.eclipse.imagen.media.range.RangeFactory;
 import org.eclipse.imagen.media.testclasses.TestBase;
+import org.eclipse.imagen.media.translate.TranslateDescriptor;
 import org.eclipse.imagen.operator.MosaicType;
 import org.eclipse.imagen.operator.NullDescriptor;
-import org.eclipse.imagen.operator.TranslateDescriptor;
 import org.junit.BeforeClass;
 import org.junit.Ignore;
 import org.junit.Test;
